refactor for easier deployment
