+++ go.mod
@@ -8,12 +8,13 @@ require (
 	github.com/containerd/continuity v0.0.0-20200710164510-efbc4488d8fe // indirect
 	github.com/docker/go-connections v0.4.0 // indirect
 	github.com/docker/go-units v0.4.0 // indirect
+	github.com/fatih/color v1.10.0
 	github.com/opencontainers/image-spec v1.0.1 // indirect
 	github.com/opencontainers/runc v0.1.1 // indirect
 	github.com/opencontainers/selinux v1.6.0 // indirect
 	github.com/pkg/errors v0.9.1
 	github.com/spf13/cast v1.3.1
+	github.com/spf13/cobra v1.1.1
 	github.com/tomochain/tomochain v0.0.0-20200528030429-718898fce456
-	go.uber.org/config v1.4.0
 	golang.org/x/sync v0.0.0-20200625203802-6e8e738ad208
 )
