+++ services/block_service.go
@@ -1,0 +1,62 @@
+// Copyright (c) 2020 TomoChain
+
+package services
+
+import (
+	"context"
+	"errors"
+	"github.com/coinbase/rosetta-sdk-go/types"
+	"github.com/tomochain/tomochain-rosetta-gateway/common"
+	"github.com/tomochain/tomochain-rosetta-gateway/configuration"
+	"github.com/tomochain/tomochain-rosetta-gateway/tomochain"
+)
+
+// BlockAPIService implements the server.BlockAPIServicer interface.
+type BlockAPIService struct {
+	config *configuration.Configuration
+	client Client
+}
+
+// NewBlockAPIService creates a new instance of a BlockAPIService.
+func NewBlockAPIService(
+	cfg *configuration.Configuration,
+	client Client,
+) *BlockAPIService {
+	return &BlockAPIService{
+		config: cfg,
+		client: client,
+	}
+}
+
+// Block implements the /block endpoint.
+func (s *BlockAPIService) Block(
+	ctx context.Context,
+	request *types.BlockRequest,
+) (*types.BlockResponse, *types.Error) {
+	if s.config.Mode != configuration.Online {
+		return nil, common.ErrUnavailableOffline
+	}
+
+	block, err := s.client.Block(ctx, request.BlockIdentifier)
+	if errors.Is(err, tomochain.ErrBlockOrphaned) {
+		return nil, common.ErrBlockOrphaned
+	}
+	if err != nil {
+		return nil, common.ErrTomoNotReady
+	}
+
+	return &types.BlockResponse{
+		Block: block,
+	}, nil
+}
+
+
+// BlockTransaction implements the /block/transaction endpoint.
+// Note: we don't implement this, since we already return all transactions
+// in the /block endpoint reponse above.
+func (s *BlockAPIService) BlockTransaction(
+	ctx context.Context,
+	request *types.BlockTransactionRequest,
+) (*types.BlockTransactionResponse, *types.Error) {
+	return nil, common.ErrNotImplemented
+}
