+++ services/mempool_service.go
@@ -1,0 +1,35 @@
+// Copyright (c) 2020 TomoChain
+
+package services
+
+import (
+	"context"
+	"github.com/coinbase/rosetta-sdk-go/server"
+	"github.com/coinbase/rosetta-sdk-go/types"
+	"github.com/tomochain/tomochain-rosetta-gateway/common"
+)
+
+// MempoolAPIService implements the server.MempoolAPIServicer interface.
+type MempoolAPIService struct {
+}
+
+// NewMempoolAPIService creates a new instance of a MempoolAPIService.
+func NewMempoolAPIService() server.MempoolAPIServicer {
+	return &MempoolAPIService{}
+}
+
+// Mempool implements the /mempool endpoint.
+func (s *MempoolAPIService) Mempool(
+	ctx context.Context,
+	request *types.NetworkRequest,
+) (*types.MempoolResponse, *types.Error) {
+	return nil, common.ErrNotImplemented
+}
+
+// MempoolTransaction implements the /mempool/transaction endpoint.
+func (s *MempoolAPIService) MempoolTransaction(
+	ctx context.Context,
+	request *types.MempoolTransactionRequest,
+) (*types.MempoolTransactionResponse, *types.Error) {
+	return nil, common.ErrNotImplemented
+}
