+++ tomochain/devnet.toml
@@ -1,0 +1,18 @@
+[Eth]
+SyncMode = "full"
+NetworkId	= 99
+
+[Node]
+DataDir = "/data"
+NoUSB = true
+HTTPHost = "0.0.0.0"
+HTTPPort = 8545
+HTTPVirtualHosts = ["*"]
+HTTPModules = ["eth", "debug", "admin"]
+IPCPath = ""
+AnnounceTxs = true
+
+[Node.P2P]
+ListenAddr = ":30311"  # flag --port
+BootstrapNodes        = ["enode://7d8ffe6d28f738d8b7c32f11fb6daa6204abae990a842025b0a969aabdda702aca95a821746332c2e618a92736538761b1660aa9defb099bc46b16db28992bc9@127.0.0.1:30301"]
+
