+++ tomochain/errors.go
@@ -1,0 +1,10 @@
+package tomochain
+import "errors"
+
+// Client errors
+var (
+	ErrBlockOrphaned         = errors.New("block orphaned")
+	ErrCallParametersInvalid = errors.New("call parameters invalid")
+	ErrCallOutputMarshal     = errors.New("call output marshal")
+	ErrCallMethodInvalid     = errors.New("call method invalid")
+)
