+++ tomochain/mainnet.toml
@@ -1,0 +1,18 @@
+[Eth]
+SyncMode = "full"
+NetworkId	= 88
+
+[Node]
+DataDir = "/data"
+NoUSB = true
+HTTPHost = "0.0.0.0"
+HTTPPort = 8545
+HTTPVirtualHosts = ["*"]
+HTTPModules = ["eth", "debug", "admin"]
+IPCPath = ""
+AnnounceTxs = true
+
+[Node.P2P]
+ListenAddr = ":30311"  # flag --port
+BootstrapNodes        = ["enode://97f0ca95a653e3c44d5df2674e19e9324ea4bf4d47a46b1d8560f3ed4ea328f725acec3fcfcb37eb11706cf07da669e9688b091f1543f89b2425700a68bc8876@3.212.20.0:30301","enode://b72927f349f3a27b789d0ca615ffe3526f361665b496c80e7cc19dace78bd94785fdadc270054ab727dbb172d9e3113694600dd31b2558dd77ad85a869032dea@188.166.207.189:30301","enode://c8f2f0643527d4efffb8cb10ef9b6da4310c5ac9f2e988a7f85363e81d42f1793f64a9aa127dbaff56b1e8011f90fe9ff57fa02a36f73220da5ff81d8b8df351@104.248.98.60:30301"]
+
